custom type mapping exmaple

+++ spring-rest2ts-examples/src/main/java/com/blueveery/springrest2ts/examples/model/ProductDTO.java
@@ -2,12 +2,12 @@ package com.blueveery.springrest2ts.examples.model;
 
 import com.blueveery.springrest2ts.examples.model.core.BaseDTO;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 
 public class ProductDTO extends BaseDTO {
     private String name;
-    private BigDecimal price;
+    private BigInteger price;
     private LocalDateTime expirationDate;
     private AddressDTO homeAddress;
 }
